refactor: move data-logging code to DataLogger object; refactor imports as necessary
