+++ packages/server/src/services/gameWorld/serverWorld.ts
@@ -11,6 +11,7 @@ import { ServerWorldDescription } from './worldMetadata';
 import { UsesRegistry } from '../../items/uses/usesRegistry';
 import { OnTickRegistry } from '../../items/on_ticks/onTickRegistry';
 import { DB } from '../database';
+import { DataLogger } from '../../grafana/dataLogger';
 
 export class ServerWorld implements GameWorld {
   private pathFinder: PathFinder;
@@ -90,25 +91,7 @@ export class ServerWorld implements GameWorld {
     FantasyDate.runTick();
 
     // log data for Grafana
-    const num_mobs = Mob.getAllMobIDs().length as number;
-    const num_items = Item.getAllItemIDs().length as number;
-    const tick_id = DB.prepare(
-      `
-        SELECT tick FROM ticks;
-      `
-    ).get() as { tick: number };
-    if (tick_id !== null) {
-      DB.prepare(
-        `
-          INSERT INTO grafanadata (tick_id, num_mobs, num_items)
-          VALUES (:tick_id, :num_mobs, :num_items);
-        `
-      ).run({
-        tick_id: tick_id.tick,
-        num_mobs: num_mobs,
-        num_items: num_items
-      });
-    }
+    DataLogger.logData();
 
     //const totalTime = Date.now() - startTime;
     //console.log('time to tick', totalTime);
